+++ src/component/progress-bar/book-progress-bar.tsx
@@ -13,16 +13,19 @@ const statusClass: Record<ReadingStatus, string> = {
 }
 
 export function BookProgressBar({ book }: Props) {
-    const progress = book.progress
+    const currentPosition = book.lastPosition
 
-    if (progress === null) {
+    if (book.status === 'finished' || currentPosition === null) {
         return <></>
     }
 
+    const percentage = currentPosition.asPercentage(book) || 0
+    const value = percentage ? `${percentage}%` : currentPosition.toString()
+
     return (
         <div className={`${styles.progressBarContainer} ${statusClass[book.status]}`}>
-            <div className={styles.progressBar} style={{ width: `${progress}%` }}></div>
-            <div className={styles.percentage}>{progress}%</div>
+            <div className={styles.progressBar} style={{ width: `${percentage}%` }}></div>
+            <div className={styles.percentage}>{value}</div>
         </div>
     )
 }