+++ src/component/table/table.tsx
@@ -40,7 +40,9 @@ export function BookTable({ books, onBookClick }: Props) {
                         <td className={styles.pubDate}>{book.metadata.published?.getFullYear() || ''}</td>
                         <td className={styles.pages}>{book.metadata.pages?.toLocaleString() || ''}</td>
                         <td className={styles.duration}>{book.metadata.duration?.toString('verbose') || ''}</td>
-                        <td>{book.progress && <BookProgressBar book={book} />}</td>
+                        <td>
+                            <BookProgressBar book={book} />
+                        </td>
                         <td>
                             <StarRating value={book.metadata.rating || 0} />
                         </td>